Implemented pause for two seconds when rovers enter a new room
Rovers keep track of their current room, and the MainController
checks if their current position is contained in their current room.
If it is not, the rover is stopped, the room that contains the position
is found, and the rovers currentRoom is updated.

Removed reward variable and functions from IControllableRover.